Add integrity check
Split synchronization logic into separate strategies.
Full strategy is the base one, it simply synchronizes given relation
from the API to local database, it also removes from local database
additional objects (if remove: true).

UpdatedSince strategy asks API only for changes since last sync for
given relation.

Check strategy performs integrity check, it simply fetches whole
relation from the API and displays objects which are missing, additional
or changed in the local database.
For changed objects ActiveModel #changes hash is displayed
