+++ lib/synced/model.rb
@@ -98,7 +98,7 @@ module Synced
     def synchronize(options = {})
       options.symbolize_keys!
       options.assert_valid_keys(:api, :fields, :include, :remote, :remove,
-        :scope)
+        :scope, :strategy)
       options[:remove]  = synced_remove unless options.has_key?(:remove)
       options[:include] = Array(synced_include) unless options.has_key?(:include)
       options[:fields]  = Array(synced_fields) unless options.has_key?(:fields)
